agent/requirements.txt
crewai>=0.152.0
crewai-tools>=0.59.0
copilotkit>=0.1.58
python-dotenv==1.0.1
uvicorn>=0.34.3
